requirements.txt
django_mailjet